fix(Solid): updates row styles
