+++ src/packages/solid/components/Row/index.ts
@@ -1,2 +1,19 @@
-export { default as default, type RowProps } from './Row.jsx';
-export { default as rowStyles, type RowStyles } from './Row.styles.js';
+/*
+ * Copyright 2023 Comcast Cable Communications Management, LLC
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+export { default as default } from './Row.jsx';
+export { default as rowStyles } from './Row.styles.js';
+export type { RowProps, RowStyles } from './Row.types.js';
